+++ androidclient/app/src/main/java/pt/ipl/isel/leic/ps/androidclient/data/api/dto/output/IngredientOutput.kt
@@ -2,5 +2,5 @@ package pt.ipl.isel.leic.ps.androidclient.data.api.dto.output
 
 class IngredientOutput(
     val identifier: Int,
-    val quantity: Int
+    val quantity: Float
 )
